requirements.txt
pytest~=6.0.2
affine~=2.3.0
numpy~=1.19.2
Shapely~=1.7.1
rasterio~=1.1.6
grpcio
grpcio-tools
google-api-core
matplotlib
setuptools~=45.2.0
imageio
geopandas~=0.8.1
pandas
tqdm~=4.49.0